requirements.txt
flask
sqlite3
functools
argon2-cffi
